requirements.txt
pvporcupine
vosk
pyaudio
openai
python-dotenv
google-cloud-texttospeech
